requirements.txt
numpy==1.22.3
Pillow==9.1.1
yapf==0.32.0
